Changed housekeeper value type in patches.
git-svn-id: svn://svn.zabbix.com/trunk@4179 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ upgrades/dbpatches/1.4/postgresql/patch/housekeeper.sql
@@ -2,7 +2,7 @@ CREATE TABLE housekeeper_tmp (
 	housekeeperid	bigint DEFAULT '0'	NOT NULL,
 	tablename	varchar(64)		DEFAULT ''	NOT NULL,
 	field		varchar(64)		DEFAULT ''	NOT NULL,
-	value		integer		DEFAULT '0'	NOT NULL,
+	value		bigint		DEFAULT '0'	NOT NULL,
 	PRIMARY KEY (housekeeperid)
 );
